[caffe2] move the SaveOp implementation from a header to a .cc file
This is a re-land of D26641600, but with the `SaveOpImpl` class marked as
`TORCH_API` to ensure that its symbols get exported properly in shared library
builds.

This moves the `SaveOp` code from `load_save_op.h` to `load_save_op.cc`.

Previously this implementation was all in the templatized `SaveOp` class, even
though most of the logic didn't depend on the template parameters.  Having
this code be in the header file slows down the build, and forces more files to
be rebuilt than necessary when changing the SaveOp code.  Having this code be
in a template class can also increase the generated code size be larger than
needed, as we don't need separate copies instantiated for each context type.

Differential Revision: [D26802576](https://our.internmc.facebook.com/intern/diff/D26802576/)

ghstack-source-id: 123045379
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53298